omit empty StaticRequestBody and make ComputedOptionalRequired a string for readability
