--- tools/codegen/codespec/computability.go
@@ -1,10 +1,0 @@
-package codespec
-
-const (
-	Computed         ComputedOptionalRequired = "computed"
-	ComputedOptional ComputedOptionalRequired = "computed_optional"
-	Optional         ComputedOptionalRequired = "optional"
-	Required         ComputedOptionalRequired = "required"
-)
-
-type ComputedOptionalRequired string
