+++ tools/codegen/models/auditing_api.yaml
@@ -7,7 +7,7 @@ schema:
           computed_optional_required: computed_optional
           tf_schema_name: audit_authorization_success
           tf_model_name: AuditAuthorizationSuccess
-          req_body_usage: 0
+          req_body_usage: all_request_bodies
           sensitive: false
           create_only: false
         - string: {}
@@ -15,7 +15,7 @@ schema:
           computed_optional_required: optional
           tf_schema_name: audit_filter
           tf_model_name: AuditFilter
-          req_body_usage: 0
+          req_body_usage: all_request_bodies
           sensitive: false
           create_only: false
         - string: {}
@@ -23,7 +23,7 @@ schema:
           computed_optional_required: computed
           tf_schema_name: configuration_type
           tf_model_name: ConfigurationType
-          req_body_usage: 3
+          req_body_usage: omit_always
           sensitive: false
           create_only: false
         - bool:
@@ -32,7 +32,7 @@ schema:
           computed_optional_required: computed_optional
           tf_schema_name: enabled
           tf_model_name: Enabled
-          req_body_usage: 0
+          req_body_usage: all_request_bodies
           sensitive: false
           create_only: false
         - string: {}
@@ -43,7 +43,7 @@ schema:
           computed_optional_required: required
           tf_schema_name: group_id
           tf_model_name: GroupId
-          req_body_usage: 3
+          req_body_usage: omit_always
           sensitive: false
           create_only: true
 operations:
@@ -54,14 +54,11 @@ operations:
     create:
         http_method: PATCH
         path: /api/atlas/v2/groups/{groupId}/auditLog
-        static_request_body: ""
     read:
         http_method: GET
         path: /api/atlas/v2/groups/{groupId}/auditLog
-        static_request_body: ""
     update:
         http_method: PATCH
         path: /api/atlas/v2/groups/{groupId}/auditLog
-        static_request_body: ""
     version_header: application/vnd.atlas.2023-01-01+json
 name: auditing_api
