+++ tools/codegen/models/custom_db_role_api.yaml
@@ -9,7 +9,7 @@ schema:
                       computed_optional_required: required
                       tf_schema_name: action
                       tf_model_name: Action
-                      req_body_usage: 0
+                      req_body_usage: all_request_bodies
                       sensitive: false
                       create_only: false
                     - list_nested:
@@ -20,7 +20,7 @@ schema:
                                   computed_optional_required: required
                                   tf_schema_name: cluster
                                   tf_model_name: Cluster
-                                  req_body_usage: 0
+                                  req_body_usage: all_request_bodies
                                   sensitive: false
                                   create_only: false
                                 - string: {}
@@ -28,7 +28,7 @@ schema:
                                   computed_optional_required: required
                                   tf_schema_name: collection
                                   tf_model_name: Collection
-                                  req_body_usage: 0
+                                  req_body_usage: all_request_bodies
                                   sensitive: false
                                   create_only: false
                                 - string: {}
@@ -36,21 +36,21 @@ schema:
                                   computed_optional_required: required
                                   tf_schema_name: db
                                   tf_model_name: Db
-                                  req_body_usage: 0
+                                  req_body_usage: all_request_bodies
                                   sensitive: false
                                   create_only: false
                       description: List of resources on which you grant the action.
                       computed_optional_required: optional
                       tf_schema_name: resources
                       tf_model_name: Resources
-                      req_body_usage: 0
+                      req_body_usage: all_request_bodies
                       sensitive: false
                       create_only: false
           description: List of the individual privilege actions that the role grants.
           computed_optional_required: optional
           tf_schema_name: actions
           tf_model_name: Actions
-          req_body_usage: 0
+          req_body_usage: all_request_bodies
           sensitive: false
           create_only: false
         - string: {}
@@ -61,7 +61,7 @@ schema:
           computed_optional_required: required
           tf_schema_name: group_id
           tf_model_name: GroupId
-          req_body_usage: 3
+          req_body_usage: omit_always
           sensitive: false
           create_only: true
         - set_nested:
@@ -72,7 +72,7 @@ schema:
                       computed_optional_required: required
                       tf_schema_name: db
                       tf_model_name: Db
-                      req_body_usage: 0
+                      req_body_usage: all_request_bodies
                       sensitive: false
                       create_only: false
                     - string: {}
@@ -80,14 +80,14 @@ schema:
                       computed_optional_required: required
                       tf_schema_name: role
                       tf_model_name: Role
-                      req_body_usage: 0
+                      req_body_usage: all_request_bodies
                       sensitive: false
                       create_only: false
           description: List of the built-in roles that this custom role inherits.
           computed_optional_required: optional
           tf_schema_name: inherited_roles
           tf_model_name: InheritedRoles
-          req_body_usage: 0
+          req_body_usage: all_request_bodies
           sensitive: false
           create_only: false
         - string: {}
@@ -95,25 +95,21 @@ schema:
           computed_optional_required: required
           tf_schema_name: role_name
           tf_model_name: RoleName
-          req_body_usage: 1
+          req_body_usage: omit_in_update_body
           sensitive: false
           create_only: true
 operations:
     delete:
         http_method: DELETE
         path: /api/atlas/v2/groups/{groupId}/customDBRoles/roles/{roleName}
-        static_request_body: ""
     create:
         http_method: POST
         path: /api/atlas/v2/groups/{groupId}/customDBRoles/roles
-        static_request_body: ""
     read:
         http_method: GET
         path: /api/atlas/v2/groups/{groupId}/customDBRoles/roles/{roleName}
-        static_request_body: ""
     update:
         http_method: PATCH
         path: /api/atlas/v2/groups/{groupId}/customDBRoles/roles/{roleName}
-        static_request_body: ""
     version_header: application/vnd.atlas.2023-01-01+json
 name: custom_db_role_api
