+++ tools/codegen/models/org_service_account_api.yaml
@@ -6,7 +6,7 @@ schema:
           computed_optional_required: computed
           tf_schema_name: client_id
           tf_model_name: ClientId
-          req_body_usage: 3
+          req_body_usage: omit_always
           sensitive: false
           create_only: false
         - string: {}
@@ -14,7 +14,7 @@ schema:
           computed_optional_required: computed
           tf_schema_name: created_at
           tf_model_name: CreatedAt
-          req_body_usage: 3
+          req_body_usage: omit_always
           sensitive: false
           create_only: false
         - string: {}
@@ -22,7 +22,7 @@ schema:
           computed_optional_required: required
           tf_schema_name: description
           tf_model_name: Description
-          req_body_usage: 0
+          req_body_usage: all_request_bodies
           sensitive: false
           create_only: false
         - string: {}
@@ -30,7 +30,7 @@ schema:
           computed_optional_required: required
           tf_schema_name: name
           tf_model_name: Name
-          req_body_usage: 0
+          req_body_usage: all_request_bodies
           sensitive: false
           create_only: false
         - string: {}
@@ -38,7 +38,7 @@ schema:
           computed_optional_required: required
           tf_schema_name: org_id
           tf_model_name: OrgId
-          req_body_usage: 3
+          req_body_usage: omit_always
           sensitive: false
           create_only: true
         - list:
@@ -47,7 +47,7 @@ schema:
           computed_optional_required: required
           tf_schema_name: roles
           tf_model_name: Roles
-          req_body_usage: 0
+          req_body_usage: all_request_bodies
           sensitive: false
           create_only: false
         - int64: {}
@@ -55,7 +55,7 @@ schema:
           computed_optional_required: required
           tf_schema_name: secret_expires_after_hours
           tf_model_name: SecretExpiresAfterHours
-          req_body_usage: 1
+          req_body_usage: omit_in_update_body
           sensitive: false
           create_only: true
         - set_nested:
@@ -66,7 +66,7 @@ schema:
                       computed_optional_required: computed
                       tf_schema_name: created_at
                       tf_model_name: CreatedAt
-                      req_body_usage: 3
+                      req_body_usage: omit_always
                       sensitive: false
                       create_only: false
                     - string: {}
@@ -74,7 +74,7 @@ schema:
                       computed_optional_required: computed
                       tf_schema_name: expires_at
                       tf_model_name: ExpiresAt
-                      req_body_usage: 3
+                      req_body_usage: omit_always
                       sensitive: false
                       create_only: false
                     - string: {}
@@ -82,7 +82,7 @@ schema:
                       computed_optional_required: computed
                       tf_schema_name: id
                       tf_model_name: Id
-                      req_body_usage: 3
+                      req_body_usage: omit_always
                       sensitive: false
                       create_only: false
                     - string: {}
@@ -90,7 +90,7 @@ schema:
                       computed_optional_required: computed
                       tf_schema_name: last_used_at
                       tf_model_name: LastUsedAt
-                      req_body_usage: 3
+                      req_body_usage: omit_always
                       sensitive: false
                       create_only: false
                     - string: {}
@@ -98,7 +98,7 @@ schema:
                       computed_optional_required: computed
                       tf_schema_name: masked_secret_value
                       tf_model_name: MaskedSecretValue
-                      req_body_usage: 3
+                      req_body_usage: omit_always
                       sensitive: false
                       create_only: false
                     - string: {}
@@ -106,32 +106,28 @@ schema:
                       computed_optional_required: computed
                       tf_schema_name: secret
                       tf_model_name: Secret
-                      req_body_usage: 3
+                      req_body_usage: omit_always
                       sensitive: true
                       create_only: false
           description: A list of secrets associated with the specified Service Account.
           computed_optional_required: computed
           tf_schema_name: secrets
           tf_model_name: Secrets
-          req_body_usage: 3
+          req_body_usage: omit_always
           sensitive: false
           create_only: false
 operations:
     delete:
         http_method: DELETE
         path: /api/atlas/v2/orgs/{orgId}/serviceAccounts/{clientId}
-        static_request_body: ""
     create:
         http_method: POST
         path: /api/atlas/v2/orgs/{orgId}/serviceAccounts
-        static_request_body: ""
     read:
         http_method: GET
         path: /api/atlas/v2/orgs/{orgId}/serviceAccounts/{clientId}
-        static_request_body: ""
     update:
         http_method: PATCH
         path: /api/atlas/v2/orgs/{orgId}/serviceAccounts/{clientId}
-        static_request_body: ""
     version_header: application/vnd.atlas.2024-08-05+json
 name: org_service_account_api
